+++ src/ajax/base.js
@@ -13,8 +13,23 @@ KISSY.add("ajax/base", function(S, JSON, Event, XhrObject) {
         },
         rnoContent = /^(?:GET|HEAD)$/,
         curLocation,
-        curLocationParts,
-        isLocal = rlocalProtocol.test(curLocationParts[1]),
+        curLocationParts;
+
+
+    try {
+        curLocation = location.href;
+    } catch(e) {
+        // Use the href attribute of an A element
+        // since IE will modify it given document.location
+        curLocation = document.createElement("a");
+        curLocation.href = "";
+        curLocation = curLocation.href;
+    }
+
+    curLocationParts = rurl.exec(curLocation);
+
+    var isLocal = rlocalProtocol.test(curLocationParts[1]),
+        transports = {},
         defaultConfig = {
             // isLocal:isLocal,
             type:"GET",
@@ -32,6 +47,9 @@ KISSY.add("ajax/base", function(S, JSON, Event, XhrObject) {
              cache: null,
              mimeType:null,
              headers: {},
+             xhrFields:{},
+             // jsonp script charset
+             scriptCharset:null,
              crossdomain:false,
              */
 
@@ -46,6 +64,7 @@ KISSY.add("ajax/base", function(S, JSON, Event, XhrObject) {
                 text:{
                     json:JSON.parse,
                     html:mirror,
+                    text:mirror,
                     xml:S.parseXML
                 }
             },
@@ -56,17 +75,7 @@ KISSY.add("ajax/base", function(S, JSON, Event, XhrObject) {
             }
         };
 
-    try {
-        curLocation = location.href;
-    } catch(e) {
-        // Use the href attribute of an A element
-        // since IE will modify it given document.location
-        curLocation = document.createElement("a");
-        curLocation.href = "";
-        curLocation = curLocation.href;
-    }
-
-    curLocationParts = rurl.exec(curLocation);
+    defaultConfig.converters.html = defaultConfig.converters.text;
 
     function setUpConfig(c) {
         c = c || {};
@@ -102,125 +111,8 @@ KISSY.add("ajax/base", function(S, JSON, Event, XhrObject) {
         return c;
     }
 
-    function fire(eventType, c, xhr) {
-        io.fire(eventType, { ajaxConfig: c ,xhr:xhr});
-    }
-
-    function handleResponseData(xhr) {
-        // text html 特殊处理
-        var text = xhr.responseText,
-            xml = xhr.responseXML;
-
-        if (!text && !xml) {
-            return;
-        }
-        var contentType = xhr.mimeType || xhr.getResponseHeader("Content-Type"),
-            c = xhr.config,
-            cConverts = c.converters,
-            xConverts = xhr.converters,
-            type,
-            responseData,
-            contents = c.contents,
-            dataType = c.dataType;
-
-        // 去除无用的通用格式
-        while (dataType[0] == "*") {
-            dataType.shift();
-        }
-
-        if (!dataType.length) {
-            // 获取源数据格式，放在第一个
-            for (type in contents) {
-                if (contents[type].test(contentType)) {
-                    if (dataType[0] != type) {
-                        dataType.unshift(type);
-                    }
-                    break;
-                }
-            }
-        }
-        // 服务器端没有告知（并且客户端没有mimetype）默认 text 类型
-        dataType[0] = dataType[0] || "text";
-
-        //获得合适的初始数据
-        if (dataType[0] == "text") {
-            responseData = text;
-        } else if (dataType[0] == "xml") {
-            responseData = xml;
-        } else {
-            // 看能否从 text xml 转换到合适数据
-            var ok = false;
-            S.each(["text","xml"], function(prevType) {
-                var type = dataType[0];
-                var converter = xConverts[prevType] && xConverts[prevType][type] ||
-                    cConverts[prevType] && cConverts[prevType][type];
-                if (converter) {
-                    responseData = converter(prevType == "text" ? text : xml);
-                    ok = true;
-                    return false;
-                }
-            });
-            if (!ok) {
-                throw " no covert for text|xml => " + dataType[0];
-            }
-        }
-
-        var prevType = dataType[0];
-
-        // 把初始数据转换成我们想要的数据类型
-        for (var i = 1; i < dataType.length; i++) {
-            type = dataType[i];
-
-            var converter = xConverts[prevType] && xConverts[prevType][type] ||
-                cConverts[prevType] && cConverts[prevType][type];
-
-            if (!converter) {
-                throw "no covert for " + prevType + " => " + type;
-            }
-            responseData = converter(responseData);
-        }
-
-        xhr.responseData = responseData;
-    }
-
-    function done(status, statusText, xhr) {
-        if (xhr.state == 2) {
-            return;
-        }
-        xhr.state = 2;
-        xhr.readyState = 4;
-        var isSuccess;
-        if (status >= 200 && status < 300 || status == 304) {
-
-            if (status == 304) {
-                statusText = "notmodified";
-                isSuccess = true;
-            } else {
-                try {
-                    handleResponseData(xhr);
-                    statusText = "success";
-                    isSuccess = true;
-                } catch(e) {
-                    statusText = "parsererror : " + e;
-                }
-            }
-
-        } else {
-            if (status < 0) {
-                status = 0;
-            }
-        }
-
-        xhr.status = status;
-        xhr.statusText = statusText;
-
-        if (isSuccess) {
-            xhr.fire("success");
-        } else {
-            xhr.fire("error");
-        }
-        xhr.fire("complete");
-        xhr.transport = undefined;
+    function fire(eventType, xhr) {
+        io.fire(eventType, { ajaxConfig: xhr.config ,xhr:xhr});
     }
 
     function handXhr(e) {
@@ -233,70 +125,69 @@ KISSY.add("ajax/base", function(S, JSON, Event, XhrObject) {
         if (c[type]) {
             c[type].call(c.context, xhr.responseData, xhr.statusText, xhr);
         }
-        fire(type, c, xhr);
+        fire(type, xhr);
     }
 
-    var transports = {},
-        io = S.mix({
-                __transports:transports,
-                ajax:function(c) {
-                    c = setUpConfig(c);
-                    var xhr = new XhrObject();
-                    fire("start", c, xhr);
-                    var transport = transports[c.dataType] && new transports[c.dataType](c);
-                    if (!transport) return;
-                    xhr.transport = transport;
+    function io(c) {
+        c = setUpConfig(c);
+        var xhr = new XhrObject(c);
+        fire("start", xhr);
+        var transportContructor = transports[c.dataType[0]] || transports["*"],
+            transport = new transportContructor(xhr);
+        xhr.transport = transport;
 
-                    if (c.contentType) {
-                        xhr.setRequestHeader("Content-Type", c.contentType);
-                    }
-                    var dataType = c.dataType,
-                        accepts = c.accepts;
-                    // Set the Accepts header for the server, depending on the dataType
-                    xhr.setRequestHeader(
-                        "Accept",
-                        dataType && accepts[dataType] ?
-                            accepts[ dataType ] + (dataType !== "*" ? ", */*; q=0.01" : "" ) :
-                            accepts[ "*" ]
-                    );
+        if (c.contentType) {
+            xhr.setRequestHeader("Content-Type", c.contentType);
+        }
+        var dataType = c.dataType[0],
+            accepts = c.accepts;
+        // Set the Accepts header for the server, depending on the dataType
+        xhr.setRequestHeader(
+            "Accept",
+            dataType && accepts[dataType] ?
+                accepts[ dataType ] + (dataType !== "*" ? ", */*; q=0.01" : "" ) :
+                accepts[ "*" ]
+        );
 
-                    // Check for headers option
-                    for (var i in c.headers) {
-                        xhr.setRequestHeader(i, c.headers[ i ]);
-                    }
+        // Check for headers option
+        for (var i in c.headers) {
+            xhr.setRequestHeader(i, c.headers[ i ]);
+        }
 
-                    xhr.on("complete success error", handXhr);
+        xhr.on("complete success error", handXhr);
 
-                    xhr.readyState = 1;
+        xhr.readyState = 1;
 
-                    fire("send", c, xhr);
+        fire("send", xhr);
 
-                    // Timeout
-                    if (c.async && c.timeout > 0) {
-                        xhr.timeoutTimer = setTimeout(function() {
-                            xhr.abort("timeout");
-                        }, c.timeout);
-                    }
+        // Timeout
+        if (c.async && c.timeout > 0) {
+            xhr.timeoutTimer = setTimeout(function() {
+                xhr.abort("timeout");
+            }, c.timeout);
+        }
 
-                    try {
-                        xhr.state = 1;
-                        transport.send(xhr, done);
-                    } catch (e) {
-                        // Propagate exception as error if not done
-                        if (xhr.status < 2) {
-                            done(-1, e, xhr);
-                            // Simply rethrow otherwise
-                        } else {
-                            S.error(e);
-                        }
-                    }
+        try {
+            xhr.state = 1;
+            transport.send();
+        } catch (e) {
+            // Propagate exception as error if not done
+            if (xhr.status < 2) {
+                xhr.callback(-1, e);
+                // Simply rethrow otherwise
+            } else {
+                S.error(e);
+            }
+        }
 
-                    return xhr;
-                }
+        return xhr;
+    }
 
-            }, Event.Target);
-
+    io.__transports = transports;
+    io.__defaultConfig = defaultConfig;
+    S.mix(io, Event.Target);
     io.isLocal = isLocal;
+
     return io;
 },
     {