+++ src/ajax/tests/ajax-json.php
@@ -1,0 +1,4 @@
+<?php
+header("Content-Type:application/json;");
+?>
+{"x":1}