+++ src/ajax/tests/ajax.php
@@ -1,0 +1,4 @@
+<?php
+header("Content-Type:text/html;");
+?>
+{"x":1}