+++ src/ajax/xhrobject.js
@@ -1,16 +1,96 @@
 /**
- * encapsulation of xhr object
+ * encapsulation of io object
+ * @author: yiminghe@gmail.com
  */
 KISSY.add("ajax/xhrobject", function(S, Event) {
 
     var // get individual response header from responseheader str
         rheaders = /^(.*?):[ \t]*([^\r\n]*)\r?$/mg;
 
-    function XhrObject() {
+    function handleResponseData(xhr) {
+
+        // text xml 是否原生转化支持
+        var text = xhr.responseText,
+            xml = xhr.responseXML;
+
+        // 例如 script 直接是js引擎执行，没有返回值，不需要自己处理
+        if (!text && !xml) {
+            return;
+        }
+        var contentType = xhr.mimeType || xhr.getResponseHeader("Content-Type"),
+            c = xhr.config,
+            cConverts = c.converters,
+            xConverts = xhr.converters || {},
+            type,
+            responseData,
+            contents = c.contents,
+            dataType = c.dataType;
+
+        // 去除无用的通用格式
+        while (dataType[0] == "*") {
+            dataType.shift();
+        }
+
+        if (!dataType.length) {
+            // 获取源数据格式，放在第一个
+            for (type in contents) {
+                if (contents[type].test(contentType)) {
+                    if (dataType[0] != type) {
+                        dataType.unshift(type);
+                    }
+                    break;
+                }
+            }
+        }
+        // 服务器端没有告知（并且客户端没有mimetype）默认 text 类型
+        dataType[0] = dataType[0] || "text";
+
+        //获得合适的初始数据
+        if (dataType[0] == "text" && text != undefined) {
+            responseData = text;
+        }
+        // 有 xml 值才直接取，否则可能还要从 xml 转
+        else if (dataType[0] == "xml" && xml != undefined) {
+            responseData = xml;
+        } else {
+            // 看能否从 text xml 转换到合适数据
+            S.each(["text","xml"], function(prevType) {
+                var type = dataType[0],
+                    converter = xConverts[prevType] && xConverts[prevType][type] ||
+                        cConverts[prevType] && cConverts[prevType][type];
+                if (converter) {
+                    dataType.unshift(prevType);
+                    responseData = prevType == "text" ? text : xml;
+                    return false;
+                }
+            });
+        }
+
+        var prevType = dataType[0];
+
+        // 按照转化链把初始数据转换成我们想要的数据类型
+        for (var i = 1; i < dataType.length; i++) {
+            type = dataType[i];
+
+            var converter = xConverts[prevType] && xConverts[prevType][type] ||
+                cConverts[prevType] && cConverts[prevType][type];
+
+            if (!converter) {
+                throw "no covert for " + prevType + " => " + type;
+            }
+            responseData = converter(responseData);
+
+            prevType = type;
+        }
+
+        xhr.responseData = responseData;
+    }
+
+    function XhrObject(c) {
         S.mix(this, {
                 // 结构化数据，如 json
                 responseData:null,
-                config:{},
+                config:c || {},
                 timeoutTimer:null,
                 responseText:null,
                 responseXML:null,
@@ -67,8 +147,50 @@ KISSY.add("ajax/xhrobject", function(S, Event) {
                 if (this.transport) {
                     this.transport.abort(statusText);
                 }
-                this.statusText = statusText;
+                this.callback(0, statusText);
                 return this;
+            },
+
+            callback:function(status, statusText) {
+                debugger
+                var xhr = this;
+                if (xhr.state == 2) {
+                    return;
+                }
+                xhr.state = 2;
+                xhr.readyState = 4;
+                var isSuccess;
+                if (status >= 200 && status < 300 || status == 304) {
+
+                    if (status == 304) {
+                        statusText = "notmodified";
+                        isSuccess = true;
+                    } else {
+                        try {
+                            handleResponseData(xhr);
+                            statusText = "success";
+                            isSuccess = true;
+                        } catch(e) {
+                            statusText = "parsererror : " + e;
+                        }
+                    }
+
+                } else {
+                    if (status < 0) {
+                        status = 0;
+                    }
+                }
+
+                xhr.status = status;
+                xhr.statusText = statusText;
+
+                if (isSuccess) {
+                    xhr.fire("success");
+                } else {
+                    xhr.fire("error");
+                }
+                xhr.fire("complete");
+                xhr.transport = undefined;
             }
         }
     );